+++ tfjs-core/src/ops/batchnorm.ts
@@ -15,18 +15,17 @@
  * =============================================================================
  */
 
-import {ENGINE} from '../engine';
+import {ENGINE, ForwardFunc} from '../engine';
+import {BatchNormalizationAttrs, BatchNormalizationInputs} from '../kernel_names';
+import {NamedAttrMap} from '../kernel_registry';
 import {Tensor, Tensor1D, Tensor4D} from '../tensor';
+import {NamedTensorMap} from '../tensor_types';
 import {convertToTensor} from '../tensor_util_env';
-import {Rank, ShapeMap, TensorLike} from '../types';
+import {Rank, TensorLike} from '../types';
 import * as util from '../util';
 
-import {warnDeprecation} from './batchnorm_util';
-import {getReductionAxes} from './broadcast_util';
+import {warnDeprecation, xAs4D} from './batchnorm_util';
 import {op} from './operation';
-import {scalar} from './tensor_ops';
-import {tile} from './tile';
-import {rsqrt} from './unary_ops';
 
 /**
  * @deprecated Please use `tf.batchNorm` instead and note the positional
@@ -100,101 +99,26 @@ function batchNorm_<R extends Rank>(
       () => 'Batch normalization gradient requires mean and scale to have ' +
           'equal ranks.');
 
-  let x4D: Tensor4D;
-  if ($x.rank === 0 || $x.rank === 1) {
-    x4D = $x.as4D(1, 1, 1, $x.size);
-  } else if ($x.rank === 2) {
-    x4D = $x.as4D(1, 1, $x.shape[0], $x.shape[1]);
-  } else if ($x.rank === 3) {
-    x4D = $x.as4D(1, $x.shape[0], $x.shape[1], $x.shape[2]);
-  } else {
-    x4D = $x as Tensor4D;
-  }
+  const x4D: Tensor4D = xAs4D($x);
 
-  const der = (dy: Tensor, saved: Tensor[]) => {
-    type Saved = [
-      Tensor<R>, Tensor<R>| Tensor1D, Tensor<R>| Tensor1D, Tensor<R>| Tensor1D
-    ];
-    const [$x, $mean, $variance, $scale] = saved as Saved;
-    const scaleValue = $scale == null ? scalar(1) : $scale;
-    const reductionAxes = getReductionAxes($mean.shape, x4D.shape);
-    const tileShape: number[] = [];
-    if ($mean.rank === 1) {
-      for (let i = 0; i < x4D.shape.length - 1; ++i) {
-        tileShape.push(x4D.shape[i]);
-      }
-      tileShape.push(1);
-    }
-
-    const xMinusMean = $x.sub($mean);
-    const dyTimesScaleValue = dy.mul(scaleValue);
-    const oneOverSqrtVariance = rsqrt($variance.add(scalar(varianceEpsilon)));
-    const minusHalfRCube = oneOverSqrtVariance.mul(oneOverSqrtVariance)
-                               .mul(oneOverSqrtVariance)
-                               .mul(scalar(-0.5));
-
-    const derX = () => {
-      if ($mean.rank === 1) {
-        return dy
-            .mul(tile(
-                oneOverSqrtVariance.as4D(1, 1, 1, $mean.shape[0]), tileShape))
-            .mul(scaleValue)
-            .reshape($x.shape);
-      } else {
-        return dy.mul(oneOverSqrtVariance).mul(scaleValue).reshape($x.shape);
-      }
-    };
-    const derMean = () => {
-      let meanDer = oneOverSqrtVariance.mul(scalar(-1)).mul(dyTimesScaleValue);
-      if ($mean.rank === 1) {
-        meanDer = meanDer.sum(reductionAxes);
-      }
-      return meanDer.reshape($mean.shape as ShapeMap[R]);
-    };
-    const derVariance = () => {
-      let varianceDer = minusHalfRCube.mul(xMinusMean).mul(dyTimesScaleValue);
-      if ($mean.rank === 1) {
-        varianceDer = varianceDer.sum(reductionAxes);
-      }
-      return varianceDer.reshape($mean.shape as ShapeMap[R]);
-    };
-    const derScale = () => {
-      const xMinusMean2TimesRsqrt = xMinusMean.mul(oneOverSqrtVariance);
-      let scaleDer = dy.mul(xMinusMean2TimesRsqrt);
-      if ($mean.rank === 1) {
-        scaleDer = scaleDer.sum(reductionAxes);
-      }
-      return scaleDer.reshape($mean.shape as ShapeMap[R]);
-    };
-    const derOffset = () => {
-      let offsetDer = dy;
-      if ($mean.rank === 1) {
-        offsetDer = offsetDer.sum(reductionAxes);
-      }
-      return offsetDer.reshape($mean.shape as ShapeMap[R]);
-    };
-    return {
-      x: derX,
-      mean: derMean,
-      variance: derVariance,
-      scale: derScale,
-      offset: derOffset
-    };
+  const forward: ForwardFunc<Tensor> = (backend, save) => {
+    const res = backend.batchNormalization(
+        x4D, batchnormReshape4D($mean), batchnormReshape4D($variance),
+        varianceEpsilon, batchnormReshape4D($scale),
+        batchnormReshape4D($offset));
+    save([$x, $mean, $variance, $scale]);
+    return res;
   };
 
-  const inputsToSave = [$x, $mean, $variance, $scale];
+  const inputs: BatchNormalizationInputs =
+      {x: $x, scale: $scale, offset: $offset, mean: $mean, variance: $variance};
+
+  const attrs: BatchNormalizationAttrs = {varianceEpsilon};
 
   const res = ENGINE.runKernelFunc(
-      (backend, save) => {
-        const res = backend.batchNormalization(
-            x4D, batchnormReshape4D($mean), batchnormReshape4D($variance),
-            varianceEpsilon, batchnormReshape4D($scale),
-            batchnormReshape4D($offset));
-        save([$x, $mean, $variance, $scale]);
-        return res;
-      },
-      {x: $x, mean: $mean, variance: $variance, scale: $scale, offset: $offset},
-      der, 'BatchNormalization', {varianceEpsilon}, inputsToSave);
+      forward, inputs as {} as NamedTensorMap, null /* gradient */,
+      'BatchNormalization', attrs as {} as NamedAttrMap);
+
   return res.reshape($x.shape);
 }
