+++ tfjs-core/src/ops/batchnorm_util.ts
@@ -15,10 +15,27 @@
  * =============================================================================
  */
 import {deprecationWarn} from '../globals';
+import {Tensor, Tensor4D} from '../tensor';
+import {Rank} from '../types';
 
-export function warnDeprecation() {
+export function warnDeprecation(): void {
   deprecationWarn(
-    'tf.batchNormalization() is going away. ' +
-    'Use tf.batchNorm() instead, and note the positional argument change ' +
-    'of scale, offset, and varianceEpsilon');
+      'tf.batchNormalization() is going away. ' +
+      'Use tf.batchNorm() instead, and note the positional argument change ' +
+      'of scale, offset, and varianceEpsilon');
+}
+
+export function xAs4D<R extends Rank>(x: Tensor<R>) {
+  let x4D: Tensor4D;
+  if (x.rank === 0 || x.rank === 1) {
+    x4D = x.as4D(1, 1, 1, x.size);
+  } else if (x.rank === 2) {
+    x4D = x.as4D(1, 1, x.shape[0], x.shape[1]);
+  } else if (x.rank === 3) {
+    x4D = x.as4D(1, x.shape[0], x.shape[1], x.shape[2]);
+  } else {
+    x4D = x as Tensor4D;
+  }
+
+  return x4D;
 }
